+++ st2actions/tests/unit/policies/test_concurrency_by_attr.py
@@ -76,6 +76,11 @@ class ConcurrencyByAttributePolicyTest(EventletTestCase, DbTestCase):
         loader.save_fixtures_to_db(fixtures_pack=PACK,
                                    fixtures_dict=TEST_FIXTURES)
 
+    def tearDown(self):
+        for liveaction in LiveAction.get_all():
+            action_service.update_status(
+                liveaction, action_constants.LIVEACTION_STATUS_CANCELED)
+
     def test_over_threshold(self):
         policy_db = Policy.get_by_ref('wolfpack.action-1.concurrency.attr')
         self.assertGreater(policy_db.parameters['threshold'], 0)
@@ -85,10 +90,8 @@ class ConcurrencyByAttributePolicyTest(EventletTestCase, DbTestCase):
             liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'fu'})
             action_service.request(liveaction)
 
-        scheduled = LiveAction.get_all()
+        scheduled = [item for item in LiveAction.get_all() if item.status in SCHEDULED_STATES]
         self.assertEqual(len(scheduled), policy_db.parameters['threshold'])
-        for liveaction in scheduled:
-            self.assertIn(liveaction.status, SCHEDULED_STATES)
 
         # Execution is expected to be delayed since concurrency threshold is reached.
         liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'fu'})
@@ -106,6 +109,38 @@ class ConcurrencyByAttributePolicyTest(EventletTestCase, DbTestCase):
         # Mark one of the execution as completed.
         action_service.update_status(
             scheduled[0], action_constants.LIVEACTION_STATUS_SUCCEEDED, publish=True)
+
+        # Execution is expected to be rescheduled.
+        liveaction = LiveAction.get_by_id(str(delayed.id))
+        self.assertIn(liveaction.status, SCHEDULED_STATES)
+
+    def test_on_cancellation(self):
+        policy_db = Policy.get_by_ref('wolfpack.action-1.concurrency.attr')
+        self.assertGreater(policy_db.parameters['threshold'], 0)
+        self.assertIn('actionstr', policy_db.parameters['attributes'])
+
+        for i in range(0, policy_db.parameters['threshold']):
+            liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'fu'})
+            action_service.request(liveaction)
+
+        scheduled = [item for item in LiveAction.get_all() if item.status in SCHEDULED_STATES]
+        self.assertEqual(len(scheduled), policy_db.parameters['threshold'])
+
+        # Execution is expected to be delayed since concurrency threshold is reached.
+        liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'fu'})
+        liveaction, _ = action_service.request(liveaction)
+        delayed = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(delayed.status, action_constants.LIVEACTION_STATUS_DELAYED)
+
+        # Execution is expected to be scheduled since concurrency threshold is not reached.
+        # The execution with actionstr "fu" is over the threshold but actionstr "bar" is not.
+        liveaction = LiveActionDB(action='wolfpack.action-1', parameters={'actionstr': 'bar'})
+        liveaction, _ = action_service.request(liveaction)
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertIn(liveaction.status, SCHEDULED_STATES)
+
+        # Cancel execution.
+        action_service.request_cancellation(scheduled[0], 'stanley')
 
         # Execution is expected to be rescheduled.
         liveaction = LiveAction.get_by_id(str(delayed.id))